requirements.txt
Flask
Flask-CORS
Flask-SocketIO
opencv-python
numpy
Pillow
torch
google-generativeai
vedo
SpeechRecognition
PyAudio
python-socketio
eventlet
